templatized the ICP class.  Wrote a mex wrapper in matlab and put together a test script.  ICP double precision is now confirmed and the performance in matlab is pretty good.

+++ include/icp/math/icp_eigen_data.h
@@ -24,8 +24,8 @@ namespace math {
     // http://eigen.tuxfamily.org/dox-devel/TopicStructHavingEigenMembers.html
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW;
 
-    Eigen::MatrixXf cross_cov_mat_;
-    Eigen::MatrixXf rot_e_mat_;
+    Eigen::MatrixXd cross_cov_mat_;
+    Eigen::MatrixXd rot_e_mat_;
   };
 
 };  // namespace math
